Many Updates
- using default msdf libary repository instead of a fork (simpler setup)
- Separate Material and Shader Manager since they require 100% separate opearations anyway
- remove "isError" parameter from most Texture functions, now they are called once when selecting a default texture, with default being "isError == false" now
- allow developer to look and decide seed inside pain::RNG class
- Application.h "setInfiniteSimulation" fnc will always detect events now
- bool isMinimized is osbolete and should be removed next commit
- PLOG macros don't need dot comma, faster allowing to type faster
- ParamSimplest from materials is now std::monostate
- Colors.h is now its own file and got separated from Misc.h
- The material component doesn't have a default material, which makes it required for the developer to pass a valid material. However, this can later be changed to add a default material if somehow the render submit fnc and the shader can be ASSERTED
- submit string can now be reversed, and its now used the modern const std::stirng_view& instead of const char *
- temporarly added the color override component to alter the color of a rendering submition without touching the material
- font manager is not a singleton anymore, it initiates together with the other renderers
- fix buggy move constructor from Font class
- default material is a standalone member, outside the std::map from the material manager
- TextBatch now normalizes the color vector, requiring 1 byte instead of 4

+++ Pain/src/CoreRender/CameraComponent.cpp
@@ -178,15 +178,16 @@ glm::vec2
 Component::OrthoCamera::screenToWorld(int mouseX, int mouseY,
                                       const pain::Transform2dComponent &camTC)
 {
-  // 1. Convert screen -> NDC space from -1 to 1
-  const float ndcX = (2.f * mouseX) / getResolution().x - 1.f;
-  const float ndcY = 1.f - (2.f * mouseY) / getResolution().y;
+  const float ndcX = (2.f * static_cast<float>(mouseX)) /
+                         static_cast<float>(getResolution().x) -
+                     1.f;
+  const float ndcY =
+      1.f -
+      (2.f * static_cast<float>(mouseY)) / static_cast<float>(getResolution().y);
 
-  // 2. Convert NDC -> camera local coordinates
   const glm::vec2 localCoord =
       glm::vec2(ndcX * m_zoomLevel * m_aspectRatio, ndcY * m_zoomLevel);
 
-  // 3. Convert camera local -> world coordinates using rotation
   return glm::vec2(camTC.m_position.x, camTC.m_position.y) + localCoord;
 }
 
@@ -195,18 +196,19 @@ Component::OrthoCamera::screenToWorld(int mouseX, int mouseY,
                                       const pain::Transform2dComponent &camTC,
                                       const pain::RotationComponent &camRC)
 {
-  // 1. Convert screen -> NDC space from -1 to 1
-  const float ndcX = (2.f * mouseX) / getResolution().x - 1.f;
-  const float ndcY = 1.f - (2.f * mouseY) / getResolution().y;
+  const float ndcX = (2.f * static_cast<float>(mouseX)) /
+                         static_cast<float>(getResolution().x) -
+                     1.f;
+  const float ndcY =
+      1.f -
+      (2.f * static_cast<float>(mouseY)) / static_cast<float>(getResolution().y);
 
-  // 2. Convert NDC -> camera local coordinates
   const glm::vec2 localCoord =
       glm::vec2(ndcX * m_zoomLevel * m_aspectRatio, ndcY * m_zoomLevel);
 
-  // 3. Convert camera local -> world coordinates using rotation
   const float angle = camRC.m_rotationRadians;
-  glm::mat2 rotation = glm::mat2(std::cos(angle), -std::sin(angle),
-                                 std::sin(angle), std::cos(angle));
+  glm::mat2 rotation = glm::mat2(std::cosf(angle), -std::sinf(angle),
+                                 std::sinf(angle), std::cosf(angle));
   return glm::vec2(camTC.m_position.x, camTC.m_position.y) +
          rotation * localCoord;
 }
